packages/pydantic-acp/pyproject.toml
[project]
name = "pydantic-acp"
version = "0.1.0"
description = "ACP adapter for pydantic-ai agents."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    # fixed for compliance with current API
    "agent-client-protocol==0.9.0",
    # fixed for compliance with current API
    "pydantic-ai-slim==1.73.0",
    "pydantic>=2.7",
    "typing-extensions>=4.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pydantic_acp"]
